requirements.txt
PyGithub
stackapi
google-generativeai
faiss-cpu
numpy
loguru